Tested creating new game, updating player attributes (we will prompt the user for this information in the later iterations)
Tested saveGame and loadGame. Made Game and Player Serializable in attempts to get ObjectOutputStream to work properly.

Added various comments throughout the code to help explain what things are doing.

Updated the BufferedReader for newRoom to include comments regarding data points.

Added Room objects to player to keep track of currentRoom and previousRoom. Established methods to facilitate setting and getting of these attributes.

May remove the contructor from Player since we are not reading in the data from a file however, we may need to switch to that approach for saving and loading rather than ObjectOutputStream.