Updated room to have monsterId, puzzleId, itemId, and npcId to facilitate adding each of these objects to their rooms.
Added methods in game to add these objects to the rooms using the respective Ids.

Tested the reader by adding data points to the end of the first line in the rooms.txt file.

Commented out the portions of code that allowed monster to keep track of all room associations - including the reader that pulled these from the file.

Added comments where appropriate.

Once the rooms.txt file is updated by TreMya, we should be good to go!