Updated Controller - moved all of the classes from game into Controller and refactored. Game only calls a new Controller and initiates the gamePlay loop.
Updated Room to allow items to be added after ReAnn completed the itemReader. Added corresponding methods.

Updated Monster to allow items to be added. Added corresponding methods.

Fixed newMonster in reader to read in item associations.

Updated the room section in the README.txt to include the additional field for the data file.